fix it's working, need more test in real condition to be sure

+++ opossum_core1/src/FEETECH/feetech.c
@@ -11,7 +11,7 @@ extern XUartPs Uart1_Instance; // from your uart code (you had this variable)
 #define RX_TIMEOUT_MS_DEFAULT 20
 
 uint8_t Com_FEETECH_Status = COM_FEETECH_IDDLE;
-uint16_t Time_Of_Last_FEETECH_Received = 0;
+uint32_t Time_Of_Last_FEETECH_Received = 0;
 uint32_t Com_FEETECH_Maxtime = 5;
 
 uint8_t FEETECH_Loop_State = 0;
